+++ .storybook/main.js
@@ -8,7 +8,6 @@ const componentPkgs = readdirSync(componentsPath, {
 	.map((dirent) => dirent.name);
 module.exports = {
 	stories: [
-		"../components/*/stories/*.stories.mdx",
 		"../components/*/stories/*.stories.@(js|jsx|ts|tsx)",
 	],
 	rootDir: "../",
@@ -33,9 +32,10 @@ module.exports = {
 		"@whitespace/storybook-addon-html",
 		// https://storybook.js.org/addons/@etchteam/storybook-addon-status
 		"@etchteam/storybook-addon-status",
-    "storybook-addon-pseudo-states",
 		// https://github.com/storybookjs/storybook/tree/next/code/addons/interactions
-		"@storybook/addon-interactions"
+		"@storybook/addon-interactions",
+		// https://www.chromatic.com/docs/visual-testing-addon/
+		"@chromaui/addon-visual-tests",
 	],
 	core: {
 		disableTelemetry: true,