+++ components/icon/stories/template.js
@@ -1,9 +1,9 @@
-import { html } from "lit";
-import { ifDefined } from "lit/directives/if-defined.js";
-import { unsafeHTML } from "lit/directives/unsafe-html.js";
+import { html, svg } from "lit";
 import { classMap } from "lit/directives/class-map.js";
+import { ifDefined } from "lit/directives/if-defined.js";
+import { unsafeSVG } from "lit/directives/unsafe-svg.js";
 
-import { fetchIconSVG, workflowIcons, uiIcons } from "./utilities.js";
+import { fetchIconSVG, uiIcons, workflowIcons } from "./utilities.js";
 
 import "../index.css";
 
@@ -93,7 +93,9 @@ export const Template = ({
 		setName = "workflow";
 	} else if (uiIcons.includes(idKey.replace(/\d{2,3}$/, ""))) {
 		setName = "ui";
-	} else if (!setName) {
+	}
+
+	if (!setName) {
 		console.warn(
 			`Icon: Could not determine the icon set for the provided icon name: ${idKey}.`
 		);
@@ -126,7 +128,7 @@ export const Template = ({
 
 	// If we found an icon above, return that value with the appended class list
 	if (icon) {
-		return html`${unsafeHTML(
+		return svg`${unsafeSVG(
 			icon.replace(
 				/^<svg(.*)>/,
 				`<svg class="${Object.entries(classList)
@@ -153,13 +155,13 @@ export const Template = ({
 			/* webpackPrefetch: true */ `!!raw-loader!@adobe/spectrum-css-workflow-icons/dist/spectrum-icons.svg`
 		);
 		import(
-			/* webpackPrefetch: true */ `!!raw-loader!@spectrum-css/icon/dist/spectrum-css-icons.svg`
+			/* webpackPrefetch: true */ `!!raw-loader!@spectrum-css/ui-icons/dist/spectrum-css-icons.svg`
 		);
 	} catch (e) {
 		console.warn(e);
 	}
 
-	return html` <svg
+	return svg`<svg
 		class=${classMap(classList)}
 		id=${ifDefined(id)}
 		style=${ifDefined(inlineStyle)}